Refactor `Filters.svelte` so we can extract `read-only` chips in "Create alert" dialog
