--- web-common/src/features/dashboards/filters/FilterContainer.svelte
@@ -1,6 +1,0 @@
-<!-- @component
-The flex container component for a set of filter chips
--->
-<div class="flex flex-row flex-wrap gap-x-2 gap-y-2 items-center">
-  <slot />
-</div>
